Resolve errors related to moving components
