+++ src/pages/Playground/Playground.js
@@ -1,6 +1,6 @@
 import NavigationBar from '../../common/NavigationBar';
 import Hiscores from './Hiscores';
-import TimePast from '../../utilities/TimePast';
+import TimePast from '../../common/utilities/TimePast';
 import { useState } from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
